+++ src/main/java/it/gov/pagopa/pu/sil/connector/send_notification/config/SendNotificationApiClientConfig.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.sil.connector.send_notification.config;
 
-import it.gov.pagopa.pu.sil.config.ApiClientConfig;
+import it.gov.pagopa.pu.sil.config.rest.ApiClientConfig;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
